requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.8.0
email-validator==2.2.0
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-dotenv
loguru==0.7.3